Skip check when build service accessed from an unrelated build
The current task may not belong to the current build tree - for instance, when a service is accessed outside the context of any task.
Concretely, this currently happens in the `GeneratePrecompiledScriptPluginAccessors` task, which launches a nested empty build and runs code that accesses a service that ofc is not declared by the "current task", as the topmost task will come from the client build instead of the nested build that actually accessed the service.